+++ src/game/rroad.h
@@ -9,7 +9,7 @@
 // asm: 	.bss	FREEZE_IT,1
 extern int FREEZE_IT;
 // asm: 	.bss	THE_TRAIN_SPEED,1	;FL
-extern int THE_TRAIN_SPEED;
+extern float THE_TRAIN_SPEED;
 // asm: 	.bss	KILL_TRAIN,1		;UD
 extern int KILL_TRAIN;
